# openapi/device-integration-api.openapi.yaml
openapi: 3.0.3
info:
  title: Device Integration API
  version: 1.0.0
  description: API for controlling Raspberry Pi connected devices.
servers:
  - url: http://localhost:3000
paths:
  /api/v1/health:
    get:
      summary: Application health
      responses:
        '200':
          description: Application is healthy
        '503':
          description: Dependency is down
  /api/v1/fan/state:
    get:
      summary: Get fan state
      responses:
        '200':
          description: Current fan state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FanState'
  /api/v1/fan/start:
    post:
      summary: Start fan
      responses:
        '202':
          description: Command accepted
        '500':
          description: Command failed
  /api/v1/fan/reset:
    post:
      summary: Reset fan to default state
      responses:
        '202':
          description: Command accepted
        '500':
          description: Command failed
  /api/v1/fan/stop:
    post:
      summary: Stop fan
      responses:
        '202':
          description: Command accepted
        '500':
          description: Command failed
  /api/v1/fan/rotate:
    post:
      summary: Toggle fan rotation
      responses:
        '202':
          description: Command accepted
        '500':
          description: Command failed
  /api/v1/fan/speed/{speed}:
    put:
      summary: Set fan speed
      parameters:
        - name: speed
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 3
      responses:
        '202':
          description: Command accepted
        '400':
          description: Invalid speed parameter
        '500':
          description: Command failed
components:
  schemas:
    FanState:
      type: object
      required:
        - isOn
        - speed
        - isRotating
      properties:
        isOn:
          type: boolean
        speed:
          type: integer
          minimum: 0
          maximum: 3
        isRotating:
          type: boolean